refactor: add action input component
