+++ website/src/components/ui/ChatInput/ChatInput.module.css
@@ -1,15 +1,20 @@
 .container {
-  display: flex;
-  align-items: flex-end;
   width: 100%;
-  gap: 16px;
+  max-width: 800px;
+  margin: 0 auto;
+}
+
+.input-wrapper {
+  position: relative;
+  width: 100%;
+  margin: 0 auto;
 }
 
 .textarea {
-  flex: 1;
+  width: 100%;
   border: none;
   border-radius: var(--radius-xl);
-  padding: 12px 20px;
+  padding: 12px 56px 12px 20px;
   outline: none;
   font-size: 1rem;
   line-height: 1.5;
@@ -18,18 +23,12 @@
   overflow: hidden;
 }
 
-.button {
+.action-button {
+  position: absolute;
+  right: 8px;
+  bottom: 8px;
   border: none;
   background: none;
   padding: 8px;
   cursor: pointer;
-}
-
-.button + .button {
-  margin-left: 16px;
-}
-
-.icon {
-  width: 36px;
-  height: 36px;
 }
